requirements.txt
pytest==7.4.3
pytest-mock==3.11.1
pytest-asyncio==0.21.1
coverage==7.3.1
discord.py==2.3.2
networkit==11.0
pygraphviz==1.11
pandas==2.1.1
matplotlib==3.8.0
pillow==10.2.0
transformers==4.35.2
torch==2.0.1
sentencepiece==0.1.99
sacremoses==0.0.35
protobuf==4.24.4
textblob==0.18.0.post0
textblob-fr==0.2.0
emoji==2.8.0
nltk==3.8.1
spacy==3.7.2
scikit-learn==1.3.0
sphinx==7.2.6
myst-parser==2.0.0
sphinx_press_theme==0.9.1
feedparser==6.0.10
